+++ src/EventEmitter.test-d.ts
@@ -2,11 +2,16 @@
 // NOTE: import from root/index to test against publicly exported types
 import { EventEmitter, SubscriptionCanceller, EventSource } from ".";
 import { expectType, expectError } from "tsd";
+import { once } from "./once";
+
+const baz = Symbol("baz");
+const bad = Symbol("bad");
 
 // Sample Events interface for testing
 interface Events {
     foo(a: number, b: boolean): void;
-    bar(): void;
+    bar(a: string): void;
+    [baz](a: boolean): void;
 }
 
 // shared instance for tests
@@ -17,7 +22,8 @@ const eventEmitter = new EventEmitter<Events>();
     // `emit` is the same as the `Events` type, except readonly
     expectType<{
         readonly foo: (a: number, b: boolean) => void;
-        readonly bar: () => void;
+        readonly bar: (a: string) => void;
+        readonly [baz]: (a: boolean) => void;
     }>(eventEmitter.emit);
 }
 
@@ -30,23 +36,42 @@ const eventEmitter = new EventEmitter<Events>();
 {
     // subscribe with invalid event name
     expectError(
-        eventEmitter.on({
+        eventEmitter.subscribe({
             bad: () => {
                 // empty
             },
         })
     );
 
-    // subscribe with partial handlers object, no options
-    eventEmitter.on({
+    // subscribe with a valid event name
+    eventEmitter.subscribe({
         foo: (a, b) => {
             expectType<number>(a);
             expectType<boolean>(b);
         },
     });
 
-    // subscribe with partial handlers object, handler returns a Promise.
-    eventEmitter.on({
+    // subscribe with multiple event names
+    eventEmitter.subscribe({
+        foo: (a, b) => {
+            expectType<number>(a);
+            expectType<boolean>(b);
+        },
+        bar: (a) => {
+            expectType<string>(a);
+        },
+    });
+
+    // subscribe using `once()`
+    eventEmitter.subscribe({
+        foo: once((a, b) => {
+            expectType<number>(a);
+            expectType<boolean>(b);
+        }),
+    });
+
+    // handler returns a Promise.
+    eventEmitter.subscribe({
         foo: (a, b) => {
             expectType<number>(a);
             expectType<boolean>(b);
@@ -54,9 +79,9 @@ const eventEmitter = new EventEmitter<Events>();
         },
     });
 
-    // subscribe with partial handlers object, handler returns an invalid type.
+    // handler returns an invalid type.
     expectError(
-        eventEmitter.on({
+        eventEmitter.subscribe({
             foo: (a, b) => {
                 expectType<number>(a);
                 expectType<boolean>(b);
@@ -65,19 +90,12 @@ const eventEmitter = new EventEmitter<Events>();
         })
     );
 
-    // subscribe with partial handlers object, partial options
-    eventEmitter.on(
-        {
-            foo: (a, b) => {
-                expectType<number>(a);
-                expectType<boolean>(b);
-            },
-        },
-        {
-            once: true,
-        }
-    );
+    // returns a SubscriptionCanceller
+    expectType<SubscriptionCanceller>(eventEmitter.subscribe({}));
+}
 
+// on()
+{
     // subscribe with invalid event name
     expectError(
         eventEmitter.on("broken", () => {
@@ -85,21 +103,21 @@ const eventEmitter = new EventEmitter<Events>();
         })
     );
 
-    // subscribe with event name, no options
+    // subscribe with valid event name
     eventEmitter.on("foo", (a, b) => {
         expectType<number>(a);
         expectType<boolean>(b);
     });
 
-    // subscribe with event name, handler returns a Promise
+    // handler returns a Promise
     eventEmitter.on("foo", (a, b) => {
         expectType<number>(a);
         expectType<boolean>(b);
         return Promise.resolve();
     });
 
+    // handler returns an invalid type
     expectError(
-        // subscribe with event name, handler returns an invalid type
         eventEmitter.on("foo", (a, b) => {
             expectType<number>(a);
             expectType<boolean>(b);
@@ -107,18 +125,59 @@ const eventEmitter = new EventEmitter<Events>();
         })
     );
 
-    // subscribe with event name, partial options
-    eventEmitter.on(
-        "foo",
-        (a, b) => {
-            expectType<number>(a);
-            expectType<boolean>(b);
-        },
-        {
-            once: true,
-        }
-    );
+    // subscribe to a valid unique eymbol event name
+    eventEmitter.on(baz, (a) => {
+        expectType<boolean>(a);
+    });
+
+    // subscribe to an invalid unique eymbol event name
+    expectError(eventEmitter.on(bad, () => undefined));
 
     // returns a SubscriptionCanceller
     expectType<SubscriptionCanceller>(eventEmitter.on("bar", () => undefined));
+}
+
+// once()
+{
+    // subscribe with invalid event name
+    expectError(
+        eventEmitter.once("broken", () => {
+            // empty
+        })
+    );
+
+    // subscribe with valid event name
+    eventEmitter.once("foo", (a, b) => {
+        expectType<number>(a);
+        expectType<boolean>(b);
+    });
+
+    // handler returns a Promise
+    eventEmitter.once("foo", (a, b) => {
+        expectType<number>(a);
+        expectType<boolean>(b);
+        return Promise.resolve();
+    });
+
+    // handler returns an invalid type
+    expectError(
+        eventEmitter.once("foo", (a, b) => {
+            expectType<number>(a);
+            expectType<boolean>(b);
+            return 42;
+        })
+    );
+
+    // subscribe to a valid unique eymbol event name
+    eventEmitter.once(baz, (a) => {
+        expectType<boolean>(a);
+    });
+
+    // subscribe to an invalid unique eymbol event name
+    expectError(eventEmitter.once(bad, () => undefined));
+
+    // returns a SubscriptionCanceller
+    expectType<SubscriptionCanceller>(
+        eventEmitter.once("bar", () => undefined)
+    );
 }