+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types.public";
 export * from "./EventEmitter";
 export * from "./WithEventEmitter";
+export { once } from "./once";